backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
gunicorn>=21.2
python-json-logger>=2.0
prometheus-client>=0.19

# OpenTelemetry
opentelemetry-api>=1.22
opentelemetry-sdk>=1.22
opentelemetry-instrumentation-django>=0.43b0
opentelemetry-instrumentation-psycopg2>=0.43b0
opentelemetry-exporter-otlp>=1.22
opentelemetry-instrumentation-logging>=0.43b0
